refactor: merge limits into pkg/config/config.go
